Add flow-regime based dissipation helper
