+++ XFoil.h
@@ -237,14 +237,6 @@ class XFoil {
   };
   EnvEnResult dampl(double hk, double th, double rt);
 
-  class DissipationResult {
-    public:
-    double di;
-    double di_hk;
-    double di_rt;
-  };
-  DissipationResult dil(double hk, double rt);
-  DissipationResult dilw(double hk, double rt);
   bool dslim(double &dstr, double thet, double msq, double hklim);
 
   bool gamqv();
